Use loadJs function to load JS files
